requirements.txt
azure-core==1.29.5
azure-identity==1.15.0
azure-storage-blob==12.19.0
prisma==0.11.0
Quart==0.19.4
